ref(profiling): Move and rename profiler class to UIProfiler
