+++ packages/browser/test/profiling/UIProfiler.test.ts
